Disable launch screen
For some reason or another Google added a launch screen
https://github.com/DirtyUnicorns/android_packages_apps_Contacts/commit/f31d69734ce90f5b773a992e841d804b2e2f06d3#diff-f3b0883918b76f5247f81e2d2f350a3d

This launch screen is seen EVERY SINGLE DAMN time you go into Contacts and it
looks off because the other AOSP apps don't havea  launch screen. So it was
either add launch screens to every other AOSP app or disable this one. This was
easier :-)

Change-Id: I2f183150ec42643af53ba6d8c6853102dcffdc00